Allow adding type providers from decomp API
